finish first

+++ src/main/java/by/piskunou/solvdlaba/web/mapper/impl/FlightRequestMapperImpl.java
@@ -1,6 +1,6 @@
 package by.piskunou.solvdlaba.web.mapper.impl;
 
-import by.piskunou.solvdlaba.domain.flights.FlightRequest;
+import by.piskunou.solvdlaba.domain.flight.FlightRequest;
 import by.piskunou.solvdlaba.domain.Passenger;
 import by.piskunou.solvdlaba.web.dto.flight.FlightRequestDTO;
 import by.piskunou.solvdlaba.web.mapper.FlightRequestMapper;
@@ -34,7 +34,7 @@ public class FlightRequestMapperImpl implements FlightRequestMapper {
 
     private List<Passenger> parsePassengers(String[] passengersAmount) {
         List<Passenger> passengers = new ArrayList<>( passengersAmount.length );
-        int j = 0;
+        byte j = 0;
         for(String n : passengersAmount) {
             for(short i = 0; i < Short.parseShort(n); i++) {
                 Passenger passenger = new Passenger( Passenger.Age.values()[j] );
